fix fix

+++ TestServer.py
@@ -897,7 +897,7 @@ def redByUctPlusDemo(ans):
     global recorder
     global Vsr_all
     print("red can move ", ans)
-    calculation_time = float(5)
+    calculation_time = float(15)
     bestp = 0
     bestm = ''
     move_dict = {'right': 0, 'down': 1, 'rightdown': 2}
